arrow_metadata_to_nif_term
